conversion: EAGLE-3 decode chunks + Mac-compat fixes
- build_eagle3_chunks.py (new): builds 4 decode mlpackages with hidden-state
  taps at fusion_layers=[8,17,34] for the EAGLE-3 draft to consume. Bypasses
  convert.py's broken monolithic path.
- test_eagle3_infer.py: Mac sanity-check fixes — apply_rope casts cos/sin to
  x.dtype (fp16 draft compat), draft cast to fp16 after load (matches target).
  Sanity pass: outputs match True, draft accept 33% on short prompt.
- build_speculative.py: HF_DIR honors GEMMA4_HF_DIR env var, falls back to
  ../output/gemma4-e2b/hf_model (previous hardcoded path pointed at a -final
  dir that convert.py no longer produces).
- docs/EAGLE3_INTEGRATION_STATE.md (new): resumable session log — artifact
  paths, environment gotchas (Python 3.11 / MPS OOM / convert.py monolithic
  bug), remaining phases (verify chunks, Swift integration, iPhone bench),
  command cheat sheet. User-requested so state survives memory churn.

Built artifacts (not committed — too large):
  output/eagle3_draft.mlpackage         210MB
  output/eagle3_fusion.mlpackage         14MB
  output/eagle3-chunks/chunk{1..4}.mlpackage  149/128/311/503 MB

+++ conversion/test_eagle3_infer.py
@@ -68,8 +68,8 @@ def build_rope_cache(max_seq, head_dim, theta, device, dtype=torch.float32):
 def apply_rope(x, cos, sin):
     half = x.shape[-1] // 2
     x1, x2 = x[..., :half], x[..., half:]
-    cos = cos[None, None, :, :]
-    sin = sin[None, None, :, :]
+    cos = cos[None, None, :, :].to(x.dtype)
+    sin = sin[None, None, :, :].to(x.dtype)
     return torch.cat([x1 * cos - x2 * sin, x2 * cos + x1 * sin], dim=-1)
 
 
@@ -294,6 +294,7 @@ def main():
     missing, unexpected = draft.load_state_dict(sd, strict=False)
     if missing:    print(f"  missing keys:    {missing[:5]} (total {len(missing)})")
     if unexpected: print(f"  unexpected keys: {unexpected[:5]} (total {len(unexpected)})")
+    draft = draft.to(torch.float16)
     draft.eval()
     n = sum(p.numel() for p in draft.parameters())
     print(f"  draft params: {n/1e6:.1f}M")
